All generic type for trackSimpleEvent
