+++ frontend/src/metabase/metadata/components/FieldSection/BehaviorSection/BehaviorSection.tsx
@@ -3,13 +3,13 @@ import { t } from "ttag";
 
 import { useGetDatabaseQuery, useUpdateFieldMutation } from "metabase/api";
 import { useMetadataToasts } from "metabase/metadata/hooks";
+import type { MetadataEditEventDetail } from "metabase/metadata/pages/shared/analytics";
 import {
   canFieldUnfoldJson,
   getRawTableFieldId,
   isFieldJsonUnfolded,
 } from "metabase/metadata/utils/field";
 import { PLUGIN_FEATURE_LEVEL_PERMISSIONS } from "metabase/plugins";
-import type { MetadataEditEventDetail } from "metabase-types/analytics";
 import type {
   DatabaseId,
   Field,
