+++ frontend/src/metabase/metadata/components/FieldSection/BehaviorSection/RemappingPicker/RemappingPicker.tsx
@@ -11,6 +11,7 @@ import {
   useUpdateFieldValuesMutation,
 } from "metabase/api";
 import { useMetadataToasts } from "metabase/metadata/hooks";
+import type { MetadataEditEventDetail } from "metabase/metadata/pages/shared/analytics";
 import { getRawTableFieldId } from "metabase/metadata/utils/field";
 import { PLUGIN_FEATURE_LEVEL_PERMISSIONS } from "metabase/plugins";
 import { FieldDataSelector } from "metabase/query_builder/components/DataSelector";
@@ -24,7 +25,6 @@ import {
   Stack,
   rem,
 } from "metabase/ui";
-import type { MetadataEditEventDetail } from "metabase-types/analytics";
 import type { Database, Field, FieldId } from "metabase-types/api";
 
 import {
