+++ frontend/src/metabase/metadata/components/FieldSection/FormattingSection/FormattingSection.tsx
@@ -3,11 +3,11 @@ import { t } from "ttag";
 
 import { useUpdateFieldMutation } from "metabase/api";
 import { useMetadataToasts } from "metabase/metadata/hooks";
+import type { MetadataEditEventDetail } from "metabase/metadata/pages/shared/analytics";
 import { getRawTableFieldId } from "metabase/metadata/utils/field";
 import { ColumnSettings } from "metabase/visualizations/components/ColumnSettings";
 import { getGlobalSettingsForColumn } from "metabase/visualizations/lib/settings/column";
 import { isCurrency } from "metabase-lib/v1/types/utils/isa";
-import type { MetadataEditEventDetail } from "metabase-types/analytics";
 import type {
   Field,
   FieldFormattingSettings as FieldSettings,
