+++ frontend/src/metabase/metadata/pages/shared/analytics.ts
@@ -1,5 +1,13 @@
 import { trackSimpleEvent } from "metabase/lib/analytics";
-import type { MetadataEditEventDetail } from "metabase-types/analytics";
+
+export type MetadataEditEventDetail =
+  | "type_casting"
+  | "semantic_type_change"
+  | "visibility_change"
+  | "filtering_change"
+  | "display_values"
+  | "json_unfolding"
+  | "formatting";
 
 export const trackMetadataChange = (detail: MetadataEditEventDetail) => {
   trackSimpleEvent({
